Move files to duke package
