+++ src/main/java/duke/InvalidInputException.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class InvalidInputException extends Exception {
 
     public InvalidInputException() {
