+++ src/main/java/duke/Parser.java
@@ -1,3 +1,4 @@
+package duke;
 public class Parser {
 
     public static Command parse(String input) {
