+++ src/main/java/duke/UnmarkCommand.java
@@ -1,14 +1,15 @@
-public class MarkCommand extends Command{
+package duke;
 
+public class UnmarkCommand extends Command{
     private final int index;
-    public MarkCommand(int i) {
+    public UnmarkCommand(int i) {
         index = i;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
-        tasks.markTask(index);
-        ui.showMarked(tasks.get(index));
+        tasks.unmarkTask(index);
+        ui.showUnmarked(tasks.get(index));
     }
 
     @Override
